HW3_Part1_22C188_SP14
I updated 1.3.  Your original solution had the last "for all" statement
as a conjunctive clause.  I changed it to an implication.  What reason
do you see for the last "for all" statement to be a conjunction when all
of the other ones are implications?  Do you agree with changing it to an
implication? Or do you have an explanation as to why it should be a
conjunctive statement?